Fixed casting warnings

+++ deflate.c
@@ -1058,7 +1058,7 @@ int32_t ZEXPORT PREFIX(deflate)(PREFIX3(stream) *strm, int32_t flush) {
         crc_finalize(s);
 #  endif
         put_uint32(s, strm->adler);
-        put_uint32(s, strm->total_in);
+        put_uint32(s, (uint32_t)strm->total_in);
     } else
 #endif
     {
